rename project & move runtime into cabal package data-file
this is much nicer than reading some random filepath at compiler
runtime. cabal should guarantee that `runtime.o` is where we expect it
to be, and the compiler fails to build if we didn't run gcc first to
product `runtime.o`.

hyphens in cabal project names are no good, so i just renamed to
camel-case.